Eric Zhang Pt 2 Additions
Here are the rest of the changes needed to complete part 2 of the project. It is a little bit messy and I realized I added everything onto the spash screen class when I should have created a separate level screen class, but we can worry about making it all nice a tidy later.