rebase task_3 with updates (#9)
* prepare task_2 for work (#2)

* created fifteen_game.h

header file for including other .cpp

* created Game_window

* Copied Game_window framework from Simple_window

* Removed Game_window to test compile

added rest of required libraries

* created Splash_screen

window with a button centered

* added instructions button and info

* changed info to Text and attached

also fixed constructor order to remove warnings

* Eric Zhang Pt 2 Additions

Here are the rest of the changes needed to complete part 2 of the project. It is a little bit messy and I realized I added everything onto the spash screen class when I should have created a separate level screen class, but we can worry about making it all nice a tidy later.

* Updated for difficulty 10 and 20

The prof sent out emails fixing the 10 and 20 puzzles so these changes update the puzzles to the correct ones

* Added .DS_Store do ignore list

* ignored DS_Store

* updated from master

* cleaned up workspace

* Add files via upload

Add a quit button, set buttons as global. I think it's better for task 4. Add the initial to start page.